arch: Split interruptions to a new module
